+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 5.4.0 (unpublished)
 
+* Escape control characters (thanks to https://github.com/rspilker).
 * Add a predicate to pretty printer to insert a space character before self-closing elements (thanks to https://github.com/rspilker).
 
 ## 5.3.0