+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 5.4.0 (unpublished)
 
+* Dart 2.16 requirement.
 * Escape control characters (thanks to https://github.com/rspilker).
 * Add a predicate to pretty printer to insert a space character before self-closing elements (thanks to https://github.com/rspilker).
 * Add predicates to normalizer to trim leading and trailing whitespaces, as well as collapse consecutive whitespaces.